I must apologise in advance for these changes.  They're my attempt at creating the parent/child relationship between machine operations and the graphics they're based on.  The original stimulus for this was to support the import of both graphics and machine operations from another something.heeks file.
The changes are far reaching and for that I appologise.  It has taken me so long to get to this point that I'm quite out of date.  I think that, with this checkin, I'm back on track.

I have added classes for Chamfer and Contour operations but no-one should get too interested yet.  Neither of them are finished.  I just decided that I needed to stop playing locally and resynchronise with the main trunk.  I figured that the TurnRough operation isn't finished either yet so this should not be too unreasonable.

If I've stuffed this up too much I won't be offended if someone reverts this change.